Add basc tabs
