+++ ground-server/src/app/controls/page.tsx
@@ -1,0 +1,11 @@
+import { auth } from "@/app/auth";
+
+export default async function Controls() {
+  const session = await auth();
+
+  if (!session) {
+    return <div>Not authenticated</div>;
+  }
+
+  return <div>Logged in - Controls</div>;
+}
