+++ ground-server/src/app/page.tsx
@@ -1,40 +1,67 @@
 "use client";
 
-import { useState } from "react";
-import Image from "next/image";
-import { SessionProvider } from "next-auth/react";
+import "react-grid-layout/css/styles.css";
+import "react-resizable/css/styles.css";
 
-import { UserAvatar } from "@/components/user-avatar";
-import { ModeToggle } from "@/components/mode-toggle";
-import { TelemetryAdder } from "@/components/telemetry-adder";
-import { Dashboard } from "@/components/dashboard/dashboard";
-import crt from "./images/crt.png";
+import React, { useState } from "react";
+import { WidthProvider, Responsive, type Layout } from "react-grid-layout";
+
+import { WidgetHandle } from "@/components/dashboard/widget-handle";
+import { DashboardWidget } from "@/components/dashboard/dashboard-widget";
 
 import { type Widget } from "@/lib/definitions";
+import { TelemetryAdder } from "@/components/telemetry-adder";
+
+const ResponsiveReactGridLayout = WidthProvider(Responsive);
 
 export default function Home() {
   const [widgets, setWidgets] = useState<Widget[]>([]);
 
+  const onLayoutChange = (layout: Layout[]) => {
+    setWidgets((prevWidgets) =>
+      prevWidgets.map((widget) => {
+        const updatedLayout = layout.find((l) => l.i === widget.layout.i);
+        if (updatedLayout) {
+          return { ...widget, layout: updatedLayout };
+        }
+        return widget;
+      })
+    );
+  };
+
+  // TODO for performance improvement, find a way that widgets does not rely on layout (then we can memoize the children and avoid re-rendering)
+  const children = widgets.map((widget) => {
+    const deleteWidget = () =>
+      setWidgets((prevWidgets) =>
+        prevWidgets.filter((w) => w.layout.i !== widget.layout.i)
+      );
+
+    return (
+      <DashboardWidget key={widget.layout.i} deleteWidget={deleteWidget}>
+        {widget.children}
+      </DashboardWidget>
+    );
+  });
+
   return (
-    <SessionProvider>
-      <nav className="flex p-4 border-b-2 items-center justify-between">
-        <Image
-          className="h-10 w-10"
-          width={100}
-          height={100}
-          src={crt}
-          alt="Cornell Rocketry Team"
-          priority={true}
-        ></Image>
-        <div className="flex gap-4">
-          <TelemetryAdder setWidgets={setWidgets} />
-          <ModeToggle />
-          <UserAvatar />
-        </div>
-      </nav>
-      <main>
-        <Dashboard widgets={widgets} setWidgets={setWidgets} />
-      </main>
-    </SessionProvider>
+    <div>
+      <ResponsiveReactGridLayout
+        className="layout"
+        draggableCancel=".drag-ignore"
+        layouts={{
+          lg: widgets.map((widget) => widget.layout),
+        }}
+        cols={{ lg: 24, md: 20, sm: 14, xs: 10, xxs: 6 }}
+        rowHeight={50}
+        autoSize={false}
+        onLayoutChange={onLayoutChange}
+        resizeHandle={<WidgetHandle />}
+      >
+        {children}
+      </ResponsiveReactGridLayout>
+      <div className="fixed bottom-10 right-10 z-50">
+        <TelemetryAdder setWidgets={setWidgets}></TelemetryAdder>
+      </div>
+    </div>
   );
 }
